requirements.txt
flask
python-dotenv
anthropic
nltk
google-generativeai
openai
